sidecar: move network-slirp-binding to cmd/sidecars
