+++ BUILD.bazel
@@ -192,7 +192,7 @@ container_bundle(
         "$(container_prefix)/$(image_prefix)example-hook-sidecar:$(container_tag)": "//cmd/sidecars/smbios:example-hook-sidecar-image",
         "$(container_prefix)/$(image_prefix)example-disk-mutation-hook-sidecar:$(container_tag)": "//cmd/sidecars/disk-mutation:example-disk-mutation-hook-sidecar-image",
         "$(container_prefix)/$(image_prefix)example-cloudinit-hook-sidecar:$(container_tag)": "//cmd/sidecars/cloudinit:example-cloudinit-hook-sidecar-image",
-        "$(container_prefix)/$(image_prefix)network-slirp-binding:$(container_tag)": "//cmd/network-slirp-binding:network-slirp-binding-image",
+        "$(container_prefix)/$(image_prefix)network-slirp-binding:$(container_tag)": "//cmd/sidecars/network-slirp-binding:network-slirp-binding-image",
         # container-disk images
         "$(container_prefix)/$(image_prefix)alpine-container-disk-demo:$(container_tag)": "//containerimages:alpine-container-disk-image",
         "$(container_prefix)/$(image_prefix)cirros-container-disk-demo:$(container_tag)": "//containerimages:cirros-container-disk-image",
@@ -333,7 +333,7 @@ container_push(
 container_push(
     name = "push-network-slirp-binding",
     format = "Docker",
-    image = "//cmd/network-slirp-binding:network-slirp-binding-image",
+    image = "//cmd/sidecars/network-slirp-binding:network-slirp-binding-image",
     registry = "$(container_prefix)",
     repository = "$(image_prefix)network-slirp-binding",
     tag = "$(container_tag)",
