+++ cmd/sidecars/network-slirp-binding/callback/callback.go
