+++ cmd/sidecars/network-slirp-binding/callback/callback_suite_test.go
