+++ cmd/sidecars/network-slirp-binding/dns/BUILD.bazel
@@ -3,7 +3,7 @@ load("@io_bazel_rules_go//go:def.bzl", "go_library")
 go_library(
     name = "go_default_library",
     srcs = ["resolvconf.go"],
-    importpath = "kubevirt.io/kubevirt/cmd/network-slirp-binding/dns",
+    importpath = "kubevirt.io/kubevirt/cmd/sidecars/network-slirp-binding/dns",
     visibility = ["//visibility:public"],
     deps = [
         "//pkg/network/dns:go_default_library",
