+++ cmd/sidecars/network-slirp-binding/domain/domain.go
