+++ cmd/sidecars/network-slirp-binding/domain/domain_suite_test.go
