+++ cmd/sidecars/network-slirp-binding/server/BUILD.bazel
@@ -3,11 +3,11 @@ load("@io_bazel_rules_go//go:def.bzl", "go_library")
 go_library(
     name = "go_default_library",
     srcs = ["server.go"],
-    importpath = "kubevirt.io/kubevirt/cmd/network-slirp-binding/server",
+    importpath = "kubevirt.io/kubevirt/cmd/sidecars/network-slirp-binding/server",
     visibility = ["//visibility:public"],
     deps = [
-        "//cmd/network-slirp-binding/callback:go_default_library",
-        "//cmd/network-slirp-binding/domain:go_default_library",
+        "//cmd/sidecars/network-slirp-binding/callback:go_default_library",
+        "//cmd/sidecars/network-slirp-binding/domain:go_default_library",
         "//pkg/hooks/info:go_default_library",
         "//pkg/hooks/v1alpha2:go_default_library",
         "//staging/src/kubevirt.io/api/core/v1:go_default_library",
